School-Management-System/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.10
astroid==2.8.5
bcrypt==3.1.7
blinker==1.4
Brlapi==0.8.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==7.1.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.5.0
distro-info==1.0
dj-database-url==0.5.0
Django==3.1.1
django-crispy-forms==1.9.2
django-heroku==0.3.1
djangorestframework==3.12.4
duplicity==0.8.20
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
get-mac==0.8.2
gpg===1.14.0-unknown
gunicorn==20.0.4
gyp==0.1
httplib2==0.18.1
idna==2.10
importlib-metadata==4.0.1
isort==5.10.1
jeepney==0.7.1
keyring==23.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lazy-object-proxy==1.6.0
libcomps==0.1.15
lockfile==0.12.2
louis==3.18.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.9
oauthlib==3.1.0
olefile==0.46
packaging==20.9
pafy==0.5.5
paramiko==2.7.2
pexpect==4.8.0
Pillow==8.1.2
platformdirs==2.4.0
protobuf==3.12.4
psutil==5.8.0
psycopg2==2.9.1
pycairo==1.16.2
pycups==2.0.1
Pygments==2.7.1
PyGObject==3.40.1
PyJWT==1.7.1
pylint==2.11.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.9.0
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.2.1
python-dateutil==2.8.1
python-debian===0.1.39ubuntu1
pytz==2019.1
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.66
requests==2.25.1
rpm==4.16.1.2
SecretStorage==3.3.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.3.0
systemd-python==234
terminator==2.1.0
toml==0.10.2
torbrowser-launcher==0.3.3
ttystatus==0.38
typing-extensions==4.0.0
ubuntu-advantage-tools==27.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.4.6+ds
wadllib==1.3.5
whitenoise==5.2.0
wrapt==1.13.3
xdg==5
xkit==0.0.0
youtube-dl==2020.12.14
zipp==1.0.0
